+++ src/openpecha/alignment/parsers/plaintext/plaintext.py
